+++ custom-recipes/image-conversion/recipe.py
@@ -2,11 +2,12 @@ from dataiku.customrecipe import get_recipe_config
 from PIL import Image
 from io import BytesIO
 import logging
-from ocr_recipes_io_utils import get_input_output
-from ocr_utils import convert_image_to_greyscale_bytes
-from ocr_utils import image_conversion_parameters
-from ocr_utils import pdf_to_pil_images_iterator
-from ocr_constants import Constants
+from text_extraction_ocr_utils.recipes_io_utils import get_input_output
+from text_extraction_ocr_utils import convert_image_to_greyscale_bytes
+from text_extraction_ocr_utils import image_conversion_parameters
+from text_extraction_ocr_utils import pdf_to_pil_images_iterator
+from text_extraction_ocr_utils import Constants
+
 
 logger = logging.getLogger(__name__)
 
@@ -22,7 +23,7 @@ for i, sample_file in enumerate(input_filenames):
     prefix = sample_file.split('.')[0]
     suffix = sample_file.split('.')[-1]
 
-    if suffix in Constants.TYPES:
+    if suffix in Constants.OCR_TYPES:
         with input_folder.get_download_stream(sample_file) as stream:
             img_bytes = stream.read()
 
@@ -39,5 +40,5 @@ for i, sample_file in enumerate(input_filenames):
         logger.info("OCR - Converted {}/{} images".format(i+1, total_images))
 
     else:
-        logger.info("OCR - Rejecting {} because it is not a {} file.".format(sample_file, '/'.join(Constants.TYPES)))
+        logger.info("OCR - Rejecting {} because it is not a {} file.".format(sample_file, '/'.join(Constants.OCR_TYPES)))
         logger.info("OCR - Rejected {}/{} images".format(i+1, total_images))