+++ tests/python/integration/test_scenario.py
@@ -9,5 +9,9 @@ def test_run_image_processing(user_dss_clients):
     dss_scenario.run(user_dss_clients, project_key=TEST_PROJECT_KEY, scenario_id="IMAGE_PROCESSING")
 
 
+def test_run_ocr(user_dss_clients):
+    dss_scenario.run(user_dss_clients, project_key=TEST_PROJECT_KEY, scenario_id="OCR")
+
+
 def test_run_text_extraction(user_dss_clients):
     dss_scenario.run(user_dss_clients, project_key=TEST_PROJECT_KEY, scenario_id="TEXT_EXTRACTION")